requirements.txt
cassio==0.1.3
datasets==2.14.6
langchain==0.0.325
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.1
